requirements.txt
altgraph==0.17.2
future==0.18.2
greenlet==1.1.2
pefile==2021.9.3
Pillow==9.1.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pywin32-ctypes==0.2.0
SQLAlchemy==1.4.35
